Don't load RE for preview panel rdef
