+++ omeroweb/webgateway/marshal.py
@@ -22,7 +22,6 @@ import omero
 import time
 import re
 import logging
-import traceback
 from future.utils import isbytes, bytes_to_native_str
 
 from omero.model.enums import PixelsTypeint8, PixelsTypeuint8, PixelsTypeint16
@@ -33,6 +32,8 @@ from omero.model.enums import PixelsTypedouble
 from omero.gateway import ChannelWrapper
 from omero.rtypes import unwrap
 from omero_marshal import get_encoder
+
+from .util import get_rendering_def, load_re
 
 logger = logging.getLogger(__name__)
 
@@ -206,39 +207,6 @@ def channelMarshal(channel):
         chan["lut"] = lut
     return chan
 
-
-def load_re(image, rsp_dict=None):
-    rsp_dict = {} if rsp_dict is None else rsp_dict
-    reOK = False
-    try:
-        reOK = image._prepareRenderingEngine()
-        if not reOK:
-            rsp_dict["Error"] = "Failed to prepare Rendering Engine for imageMarshal"
-            logger.debug("Failed to prepare Rendering Engine for imageMarshal")
-    except omero.ConcurrencyException as ce:
-        backOff = ce.backOff
-        rsp_dict["ConcurrencyException"] = {"backOff": backOff}
-    except Exception as ex:  # Handle everything else.
-        rsp_dict["Exception"] = ex.message
-        logger.error(traceback.format_exc())
-    return reOK
-
-
-def get_rendering_def(image, rsp_dict):
-    # Try to get user's rendering def
-    exp_id = image._conn.getUserId()
-    rdefs = image.getAllRenderingDefs(exp_id)
-    if len(rdefs) == 0:
-        # try to create our own rendering settings
-        if load_re(image, rsp_dict):
-            rdefs = image.getAllRenderingDefs(exp_id)
-    # otherwise use owners
-    if len(rdefs) == 0:
-        owner_id = image.getDetails().getOwner().id
-        if owner_id != exp_id:
-            rdefs = image.getAllRenderingDefs(owner_id)
-
-    return rdefs[0] if len(rdefs) > 0 else None
 
 
 def imageMarshal(image, key=None, request=None):
